% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW_DunnTest_pvals.R
\name{KW_DunnTest_pvals}
\alias{KW_DunnTest_pvals}
\title{StatTest and Posthoc P values of multiple interdependent variables in a
dataframe (based on Rstatix)}
\usage{
KW_DunnTest_pvals(x, VC1_col, VC2_col, p.adj.method)
}
\value{
Data frame of p values
}
\description{
StatTest and Posthoc P values of multiple interdependent variables in a
dataframe (based on Rstatix)
}
